Fixups of authorisation stuff so the right roles can see/do what they need. Added first test - fixed a bug on User->availableProjects() as a result

+++ projects2/app/Http/Controllers/CourseController.php
@@ -21,6 +21,7 @@ class CourseController extends Controller
      */
     public function index()
     {
+        $this->authorize('edit_courses');
         $courses = Course::orderBy('code')->get();
         return view('course.index', compact('courses'));
     }
@@ -32,6 +33,7 @@ class CourseController extends Controller
      */
     public function create()
     {
+        $this->authorize('edit_courses');
         $course = new Course;
         return view('course.create', compact('course'));
     }
@@ -44,6 +46,7 @@ class CourseController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('edit_courses');
         $this->validate($request, [
             'code' => 'required|unique:courses',
             'title' => 'required'
@@ -63,6 +66,7 @@ class CourseController extends Controller
      */
     public function show($id)
     {
+        $this->authorize('edit_courses');
         $course = Course::findOrFail($id);
         return view('course.show', compact('course'));
     }
@@ -75,6 +79,7 @@ class CourseController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('edit_courses');
         $course = Course::findOrFail($id);
         return view('course.edit', compact('course'));
     }
@@ -88,6 +93,11 @@ class CourseController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('edit_courses');
+        $this->validate($request, [
+            'code' => 'required|unique:courses,code,' . $id,
+            'title' => 'required'
+        ]);
         $course = Course::findOrFail($id);
         $course->fill($request->input());
         $course->save();
@@ -103,6 +113,7 @@ class CourseController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('edit_courses');
         $course = Course::findOrFail($id);
         EventLog::log(Auth::user()->id, "Deleted course {$course->title} {$course->code}");
         $course->delete();
@@ -111,12 +122,14 @@ class CourseController extends Controller
 
     public function editStudents($id)
     {
+        $this->authorize('edit_courses');
         $course = Course::findOrFail($id);
         return view('course.edit_students', compact('course'));
     }
 
     public function updateStudents(Request $request, $id)
     {
+        $this->authorize('edit_courses');
         $course = Course::findOrFail($id);
         $studentIds = [];
         $sheet = Excel::load($request->file('file'))->get();
@@ -173,6 +186,7 @@ class CourseController extends Controller
 
     public function removeStudents($courseId)
     {
+        $this->authorize('edit_courses');
         $course = Course::findOrFail($courseId);
         foreach ($course->students as $student) {
             // Note: this will remove any allocations to projects too
